phd_financial_assistant/requirements.txt
alpaca-trade-api
streamlit
yfinance
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
pandas-datareader
matplotlib
seaborn
plotly
statsmodels
scipy
beautifulsoup4
lxml
html5lib
tweepy
nltk
textblob
vaderSentiment
feedparser